[platform-api] overriding default weak otPlatTimeGet
